fix(hosts_helper): fix legacy ui
